Generic core_file_merger and gdb fixes
Several parts of the core_file_merger script had to
be tweaked to make it run on Ubuntu 20.04 + python3.8
without any fatal errors. Unclear if it affects other
debugger backends - seems so.

In the gdb script on resume function, after sending
continue command buffer is flushed to prevent deadlock.
For serial attach scenario deadlock was inevitable on
Ubuntu 20.04 + python3.8.

Co-authored-by: Oguzhan Karakaya <oguzhan.karakaya@intel.com>