Remove STAT_COLLECT_SIMD_BT from gdb-oneapi support
The "STAT_COLLECT_SIMD_BT" option in gdb-oneapi STAT support is used to add
lane information in thread id.  So, using this environment increases the
STAT capture time due to sequential calls per lane of a thread for the
collection of the call stack.  It is removed in this change because the
function call stacks collected for lanes are identical , except for
function argument values, which are not taken into consideration by STAT
anyway.

Signed-off-by: Abdul Basit Ijaz <abdul.b.ijaz@intel.com>