Improvements to gdb-oneapi support (#53)
* Improvement in getting threads list for the gdb-oneapi support

Replaced the MI command "-thread-info" with "thread apply" for retrieving
the thread list.  The previous MI command output could grow excessively
large, making it inefficient for the readlines function to process.  Now
after this change the output of "thread apply" command is only a list
of threads and will take less time in readlines function.

The test file oneapi_gdb_test.py has been updated to test these changes.

Signed-off-by: Abdul Basit Ijaz <abdul.b.ijaz@intel.com>

* Improvements in getting backtrace for the gdb-oneapi support

The backtrace command now uses the "with print frame-arguments none --"
option to enhance performance, as frame argument details are unnecessary
for the STAT tool.

Signed-off-by: Abdul Basit Ijaz <abdul.b.ijaz@intel.com>

* Remove STAT_COLLECT_SIMD_BT from gdb-oneapi support

The "STAT_COLLECT_SIMD_BT" option in gdb-oneapi STAT support is used to add
lane information in thread id.  So, using this environment increases the
STAT capture time due to sequential calls per lane of a thread for the
collection of the call stack.  It is removed in this change because the
function call stacks collected for lanes are identical , except for
function argument values, which are not taken into consideration by STAT
anyway.

Signed-off-by: Abdul Basit Ijaz <abdul.b.ijaz@intel.com>

---------

Signed-off-by: Abdul Basit Ijaz <abdul.b.ijaz@intel.com>